Fix a few things

+++ tests/WangSintering/test2d.py
@@ -3,24 +3,24 @@ import sys
 import subprocess
 import os
 
-print("Test Two grains with Quadratic energies...")
+print("Test Two grains sintering...")
 
 mpicmd  = sys.argv[1]+" "+sys.argv[2]+" "+sys.argv[3]
 exe     = sys.argv[4]
 inp     = sys.argv[5]
 datadir = sys.argv[6]
 
-print(datadir)
 #make symbolic link to calphad data
 data = "2spheres.csv"
-src = datadir+'/'+data
-os.symlink(src, data)
+if not os.path.exists(data):
+  src = datadir+'/'+data
+  os.symlink(src, data)
 
 #prepare initial conditions file
 initfilename="2spheres.nc"
 subprocess.call(["python3", "../../utils/make_multi_spheres.py",
-  "--nx", "64", "--ny", "64", "--nz", "1",
-  "--concentration-A", "0.1", "--concentration-out", "0.06",
+  "--nx", "120", "--ny", "200", "--nz", "1",
+  "--concentration-A", "1.,0.", "--concentration-B", "0.,1.", "--concentration-out", "0.,0.",
   "--spheres", data,
   initfilename])
 
@@ -36,19 +36,22 @@ lines=output.split(b'\n')
 volfractions=[]
 
 end_reached = False
+end_time = 0.006
 for line in lines:
-  if (line.count(b'phase 0') or line.count(b'phase 1') ) and line.count(b'Volume'):
-    print(line)
-    words=line.split()
-    volume=eval(words[6])
-    volfractions.append(volume)
 
   if line.count(b'cycle'):
     print(line)
     words=line.split()
     time=eval(words[6])
-    if time>0.25:
+    if time>end_time:
       end_reached = True
+
+  if end_reached:
+    if line.count(b'phase') and line.count(b'Volume'):
+      print(line)
+      words=line.split()
+      volume=eval(words[6])
+      volfractions.append(volume)
 
 minv=1.
 maxv=0.
@@ -58,15 +61,16 @@ for v in volfractions:
   if v>maxv:
     maxv = v
 
-expected_value=0.235
+expected_value=0.586
 if abs(maxv-expected_value)>0.001:
   print("Expected maxv = {}, found {}".format(expected_value,maxv))
   sys.exit(1)
 
-expected_value=0.040
+expected_value=0.118
 if abs(minv-expected_value)>0.001:
   print("Expected minv = {}, found {}".format(expected_value,minv))
   sys.exit(1)
+
 
 if end_reached:
   sys.exit(0)
